Add Today Layout. Update weather in Today Extension.

+++ WeatherInfoKit/WeatherService.swift
@@ -52,7 +52,7 @@ public class WeatherService {
                 }
                 
                 let jsonMain = jsonResult?["main"] as! Dictionary<String, AnyObject>
-                weatherData.temperature = jsonMain["temp"] as! Int
+                weatherData.temperature = Int(jsonMain["temp"] as! Double)
                 
                 completion(weatherData)
